Update README.md
With tremendous increase in technology, information technology is a fast-developing field. 
Technology which is in vogue today might become redundant tomorrow. The ever-changing
scenario makes it possible to provide the latest and most modern IT solutions to various business
and institutions. The application is for Restaurants. It has corporate functions which encompasses 
the processes for the management of recipes used by the fresh departments in the stores. This 
includes the processes for creation, approval, maintenance and deletion of recipes. The recipes are 
used by replenishment planning processes to determine ingredient and supply item requirement 
and plan the replenishment of the same at the stores.
Thus, the project is a sincere effort in simplifying the task of administrators in an easily usable 
format. This Recipe Management System has the solution to manage, update, approve, and search 
for the recipes. The system uses Tkinter of Python as the front-end development language, and
MySQL as the database for the recipes.